requirements.txt
flask==3.0.2
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==1.0.1
Pillow==10.2.0
requests==2.31.0
flask-cors==4.0.0 